+++ insight-brain-service/src/main/java/com/sonatype/insight/brain/telemetry/PullRequestCommentTelemetry.java
@@ -22,7 +22,7 @@ public class PullRequestCommentTelemetry
 
   public int prNumber;
 
-  public int commentId;
+  public long commentId;
 
   public String action;
